Added python requirements.txt
I'm a little unused to the style of this project, but I was surprised to see so many advocating for distro-maintained capstone dependency

This allows people like me that use `pip` and `virtualenv` to not add python dependencies to their base machine. It should be noted that `libcapstone-dev` and C deps are still required.